+++ web/eslint.config.mjs
@@ -81,6 +81,7 @@ export default [
       'unicorn/prevent-abbreviations': 'off',
       'unicorn/no-nested-ternary': 'off',
       'unicorn/consistent-function-scoping': 'off',
+      'unicorn/filename-case': 'off',
       'unicorn/prefer-top-level-await': 'off',
       'unicorn/import-style': 'off',
       'svelte/button-has-type': 'error',
