+++ web/src/app.html
@@ -101,6 +101,12 @@
     <link rel="stylesheet" href="/custom.css" />
   </head>
 
+  <noscript
+    class="absolute z-[1000] flex h-screen w-screen place-content-center place-items-center bg-immich-bg dark:bg-immich-dark-bg dark:text-immich-dark-fg"
+  >
+    To use Immich, you must enable JavaScript or use a JavaScript compatible browser.
+  </noscript>
+
   <body class="bg-immich-bg dark:bg-immich-dark-bg">
     <div id="stencil">
       <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 792 792">
